modularizes the favorite routes into the favoriteRoutes.js file
