+++ views/layouts/main.handlebars
@@ -27,7 +27,6 @@
     {{#if logged_in}}
         <script src="/js/logout.js"></script>
     {{/if}}
-    {{!-- <script src="/js/utilities.js"></script> --}}
     <script src="/js/main.js"></script>
 
 </body>
